fix(macos): split provider wizard setup flow
